- change location of user danger zone - add auth0 management vars to env example
