+++ .env.example
@@ -4,7 +4,8 @@ AUTH0_CLIENT_ID=xxxxx
 AUTH0_CLIENT_SECRET=xxxxx
 AUTH0_DOMAIN="sesf.eu.auth0.com"
 AUTH0_CALLBACK_URL="http://localhost:3000/auth/callback"
-CLOUDFLARE_BEARER=xxxxx
+AUTH0_MANAGEMENT_CLIENT_ID=xxxxxx
+AUTH0_MANAGEMENT_CLIENT_SECRET=xxxxxxx
 MINIO_ENDPOINT=xxxxx
 MINIO_PORT=xxxxx
 MINIO_ACCESS_KEY=xxxxx
