+++ app/routes/admin+/user+/account.tsx
@@ -1,14 +1,20 @@
-import styles from 'react-image-crop/dist/ReactCrop.css?url'
-import dateInputStyles from "~/styles/date-input.css?url";
 import type { ActionFunctionArgs, LoaderFunctionArgs } from "@remix-run/node";
-import { checkUserAuth } from "~/utils/auth.server";
-import { json, useLoaderData } from "@remix-run/react";
+import { json, useFetcher, useLoaderData } from "@remix-run/react";
+import { useState } from 'react';
+import styles from 'react-image-crop/dist/ReactCrop.css?url';
+import { z } from "zod";
+import { zx } from 'zodix';
+import EntityDetailBlock from "~/components/Blocks/EntityDetailBlock";
+import ActionButton from '~/components/Button/ActionButton';
+import AskModalBody from '~/components/Notifications/AskModalBody';
+import Modal from '~/components/Notifications/Modal';
+import H1 from '~/components/Titles/H1';
+import { entityToPathSegment } from '~/helpers/entityType';
 import { db } from "~/services/db.server";
 import { getSearchParams } from "~/services/search.server";
-import { zx } from 'zodix';
-import { z } from "zod";
-import EntityDetailBlock from "~/components/Blocks/EntityDetailBlock";
 import { createFlashMessage } from "~/services/toast.server";
+import dateInputStyles from "~/styles/date-input.css?url";
+import { checkUserAuth } from "~/utils/auth.server";
 
 export function links() {
   return [
@@ -96,6 +102,37 @@ export async function loader({ request }: LoaderFunctionArgs) {
 
 export default function () {
   const { user, searchParams } = useLoaderData<typeof loader>();
-  return <EntityDetailBlock {...user} entityId={user.id} entityType='USER' entityBirthday={user.birth_date}
-    imageId={user.image} searchParams={searchParams} />
+  const [modalOpen, setModalOpen] = useState(false);
+  const fetcher = useFetcher();
+  const handleDelete = () => {
+    setModalOpen(false);
+    const path = entityToPathSegment('USER');
+    fetcher.submit({
+      entityId: user.id.toString(),
+    }, {
+      method: 'delete',
+      action: `/admin/api/${path}`,
+    });
+  };
+  return <>
+    <EntityDetailBlock {...user} entityId={user.id} entityType='USER' entityBirthday={user.birth_date}
+      imageId={user.image} searchParams={searchParams} />
+    <div className="bg-red-600/25 py-8 lg:pb-12 my-8 px-5">
+      <div className="w-full max-w-prose mx-auto">
+        <H1>Danger Zone</H1>
+        <div className="flex flex-col items-center max-w-md mx-auto mt-8 gap-4">
+          <ActionButton content="Change Password" action={() => fetcher.submit({}, {
+            action: '/admin/api/password',
+            method: 'post',
+          })} />
+          <ActionButton content="Delete" action={() => setModalOpen(true)} />
+        </div>
+      </div>
+    </div>
+    <Modal isOpen={modalOpen} handleClose={() => setModalOpen(false)}>
+      <AskModalBody message={`Do you really want to delete your account?`}
+        primaryButton={{ text: 'Yes', onClick: handleDelete }}
+        secondaryButton={{ text: 'No', onClick: () => setModalOpen(false) }} />
+    </Modal>
+  </>
 }
